fix: resolved https://github.com/LiterateInk/Pawdirecte/pull/4#discussion_r1779756948 & https://github.com/LiterateInk/Pawdirecte/pull/4#discussion_r1779756905
