+++ src/models/index.ts
@@ -18,6 +18,7 @@ export * from "./files";
 export * from "./grade";
 export * from "./grade-kind";
 export * from "./grade-value";
+export * from "./grades-overview";
 export * from "./homepage-timeline-item";
 export * from "./homework";
 export * from "./period";
